+++ src/Vulkan/LLGI.SingleFrameMemoryPoolVulkan.cpp
@@ -102,7 +102,7 @@ SingleFrameMemoryPoolVulkan::SingleFrameMemoryPoolVulkan(
 		SafeAddRef(graphics_);
 	}
 
-	for (size_t i = 0; i < swapBufferCount; i++)
+	for (int32_t i = 0; i < swapBufferCount; i++)
 	{
 		auto memoryPool = std::make_shared<InternalSingleFrameMemoryPoolVulkan>();
 		if (!memoryPool->Initialize(graphics, constantBufferPoolSize, drawingCount))
